@@ -3,21 +3,21 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@material-ui/core": "^4.7.0",
-        "@material-ui/icons": "^4.5.1",
-        "@types/jest": "24.0.23",
-        "@types/node": "12.12.12",
-        "@types/react": "^16.9.13",
-        "@types/react-dom": "16.9.4",
-        "@types/react-router-dom": "^5.1.3",
-        "@types/swagger-ui-react": "^3.23.0",
-        "prettier": "^1.19.1",
-        "react": "^16.12.0",
-        "react-dom": "^16.12.0",
-        "react-router-dom": "^5.1.2",
-        "react-scripts": "3.2.0",
-        "swagger-ui-react": "^3.24.3",
-        "typescript": "3.7.2"
+        "@emotion/react": "^11.10.5",
+        "@emotion/styled": "^11.10.5",
+        "@mui/material": "^5.10.16",
+        "@types/jest": "^29.2.3",
+        "@types/node": "^18.11.10",
+        "@types/react": "^18.0.25",
+        "@types/react-dom": "^18.0.9",
+        "@types/react-router-dom": "^5.3.3",
+        "@types/swagger-ui-react": "^4.11.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-router-dom": "^6.4.4",
+        "react-scripts": "^5.0.1",
+        "swagger-ui-react": "^4.15.5",
+        "typescript": "^4.9.3"
     },
     "prettier": {
         "trailingComma": "es5",
@@ -45,5 +45,9 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
+    },
+    "devDependencies": {
+        "prettier": "^2.8.0",
+        "yarn-upgrade-all": "^0.7.1"
     }
 }
