@@ -1,26 +1,16 @@
 import React from 'react'
-import { makeStyles, createStyles, Theme } from '@material-ui/core/styles'
-import Grid from '@material-ui/core/Grid'
+import { Grid } from '@mui/material'
 
 import Widget from './Widget'
-
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            flexGrow: 1,
-        },
-    })
-)
 
 type Props = {
     widgets: string[]
 }
 
 export default function WidgetsGrid(props: Props) {
-    const classes = useStyles()
 
     return (
-        <div className={classes.root}>
+        <div>
             <Grid container spacing={3}>
                 {props.widgets.map(widget => (
                     <Grid key={widget} item xs>
