@@ -1,121 +1,13 @@
 import React from 'react'
-import AppBar from '@material-ui/core/AppBar'
-import Toolbar from '@material-ui/core/Toolbar'
-import Typography from '@material-ui/core/Typography'
-import InputBase from '@material-ui/core/InputBase'
-import { createStyles, fade, Theme, makeStyles } from '@material-ui/core/styles'
-import Link from '@material-ui/core/Link'
-import OpenInNew from '@material-ui/icons/OpenInNew'
-import SearchIcon from '@material-ui/icons/Search'
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            flexGrow: 1,
-        },
-        menuButton: {
-            marginRight: theme.spacing(2),
-        },
-        title: {
-            flexGrow: 1,
-            display: 'none',
-            [theme.breakpoints.up('sm')]: {
-                display: 'block',
-            },
-        },
-        search: {
-            position: 'relative',
-            borderRadius: theme.shape.borderRadius,
-            backgroundColor: fade(theme.palette.common.white, 0.15),
-            '&:hover': {
-                backgroundColor: fade(theme.palette.common.white, 0.25),
-            },
-            marginLeft: 0,
-            width: '100%',
-            [theme.breakpoints.up('sm')]: {
-                marginLeft: theme.spacing(1),
-                width: 'auto',
-            },
-        },
-        searchIcon: {
-            width: theme.spacing(7),
-            height: '100%',
-            position: 'absolute',
-            pointerEvents: 'none',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-        },
-        swagger: {
-            color: '#fff',
-        },
-        link: {
-            display: 'flex',
-            flexDirection: 'row',
-            alignItems: 'center',
-        },
-        inputRoot: {
-            color: 'inherit',
-        },
-        inputInput: {
-            padding: theme.spacing(1, 1, 1, 7),
-            transition: theme.transitions.create('width'),
-            width: '100%',
-            [theme.breakpoints.up('sm')]: {
-                width: 120,
-                '&:focus': {
-                    width: 200,
-                },
-            },
-        },
-    })
-)
 
 type Props = {
     onSearchTextChange: (word: string) => any
 }
 
 export default function SearchAppBar(props: Props) {
-    const classes = useStyles()
 
     return (
-        <div className={classes.root}>
-            <AppBar position="static">
-                <Toolbar>
-                    <Typography className={classes.title} variant="h6" noWrap>
-                        EDCS Tester
-                    </Typography>
-                    <Typography className={classes.swagger}>
-                        <Link
-                            className={classes.link}
-                            rel="noopener"
-                            target="_blank"
-                            href="/apidocs"
-                            color="inherit"
-                            variant="inherit"
-                        >
-                            <OpenInNew />
-                            Swagger UI
-                        </Link>
-                    </Typography>
-                    <div className={classes.search}>
-                        <div className={classes.searchIcon}>
-                            <SearchIcon />
-                        </div>
-                        <InputBase
-                            placeholder="Search…"
-                            classes={{
-                                root: classes.inputRoot,
-                                input: classes.inputInput,
-                            }}
-                            onChange={event =>
-                                props.onSearchTextChange(event.target.value)
-                            }
-                            inputProps={{ 'aria-label': 'search' }}
-                        />
-                    </div>
-                </Toolbar>
-            </AppBar>
-        </div>
+       <></>
     )
 }
